added limiter

+++ MoneyControl.WebAPI.Application/Services/Models/FullPaymentModel.cs
@@ -5,7 +5,7 @@
         public string UserFirstName { get; set; }
         public string UserLastName { get; set; }
         public DateTime WhenSpend { get; set; }
-        public decimal MoneySpend { get; set; }
+        public double MoneySpend { get; set; }
         public string ExpensesType { get; set; }
     }
 }
